feat: remocao da estrutura de clientes, solucao foi movida para lambda aws
