--- app/src/main/java/br/com/appfastfood/cliente/aplicacao/adaptadores/requisicao/RequisicaoExcecao.java
@@ -1,20 +1,0 @@
-package br.com.appfastfood.cliente.aplicacao.adaptadores.requisicao;
-
-public class RequisicaoExcecao {
-
-
-    private String mensagem;
-    private int codigo;
-    public RequisicaoExcecao(String mensagem, int status){
-        this.mensagem = mensagem;
-        this.codigo = status;
-    }
-
-    public String getMensagem() {
-        return mensagem;
-    }
-
-    public int getCodigo() {
-        return codigo;
-    }
-}
