--- app/src/main/java/br/com/appfastfood/cliente/infraestrutura/SpringDataClienteRepository.java
@@ -1,9 +1,0 @@
-package br.com.appfastfood.cliente.infraestrutura;
-
-import br.com.appfastfood.cliente.infraestrutura.entidades.EntidadeCliente;
-import org.springframework.data.mongodb.repository.MongoRepository;
-
-import java.util.UUID;
-
-public interface SpringDataClienteRepository extends MongoRepository<EntidadeCliente, UUID> {
-}
