--- app/src/main/java/br/com/appfastfood/cliente/usecase/portas/ClienteServico.java
@@ -1,12 +1,0 @@
-package br.com.appfastfood.cliente.usecase.portas;
-
-
-import br.com.appfastfood.cliente.dominio.modelos.Cliente;
-
-import java.util.UUID;
-
-public interface ClienteServico {
-    UUID cadastrar(String nome, String cpf, String email);
-    Cliente buscarPorCpf(String cpf);
-
-}
